chore: keep runtime package file baseline
